+++ coroutine.c
@@ -112,25 +112,8 @@ static void coroutine_object_destroy(zend_object *object)
 	}
 
 	if (coroutine->coroutine.fcall) {
-
-		zend_fcall_t *fcall = coroutine->coroutine.fcall;
+		zend_fcall_release(coroutine->coroutine.fcall);
 		coroutine->coroutine.fcall = NULL;
-
-		if (fcall->fci.param_count) {
-			for (uint32_t i = 0; i < fcall->fci.param_count; i++) {
-				zval_ptr_dtor(&fcall->fci.params[i]);
-			}
-
-			efree(fcall->fci.params);
-		}
-
-		if (fcall->fci.named_params) {
-			GC_DELREF(fcall->fci.named_params);
-			fcall->fci.named_params = NULL;
-		}
-
-		zval_ptr_dtor(&fcall->fci.function_name);
-		efree(fcall);
 	}
 
 	if (coroutine->coroutine.context != NULL) {
@@ -749,6 +732,8 @@ bool async_coroutine_resume(zend_coroutine_t *coroutine, zend_object *error, con
 	// Add to resumed_coroutines queue for event cleanup
 	if (in_scheduler_context) {
 		circular_buffer_push_ptr_with_resize(&ASYNC_G(resumed_coroutines), coroutine);
+	} else if (coroutine->waker != NULL) {
+		ZEND_ASYNC_WAKER_CLEAN_EVENTS(coroutine->waker);
 	}
 
 	return true;